api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy>=2.0
pymysql>=1.1
cryptography>=44.0
alembic>=1.14
pydantic>=2.10
pydantic-settings>=2.7
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
httpx>=0.28
apscheduler>=3.10
python-dotenv>=1.0
python-multipart>=0.0.19
